bs suspense thing
